Function modification

+++ Functions/src/exploreDune.kt
@@ -1,10 +1,10 @@
-fun exploreDune ( height: Double, climeRate: Double ): Double {
+fun exploreDune ( height: Int, climeRate: Int ): Double {
 
-    return height/ climeRate
+    return height/ climeRate.toDouble()
 }
 
 fun main () {
 
-    println(" Time to Climb the Dune: =" +  exploreDune( 20.00, 3.0) + " hours")
+    println(" Time to Climb the Dune: =" +  exploreDune( 20, 3) + " hours")
 //println (exploreDune(20.0, 3.0))
 }
